Move server-side eval function to server.py
