+++ baselines/fjord/fjord/main.py
@@ -3,7 +3,7 @@ import math
 import os
 import random
 from types import SimpleNamespace
-from typing import Any, Callable, Dict, List, Optional, Tuple, Union
+from typing import Any, Callable, Dict, List, Optional, Union
 
 import flwr as fl
 import hydra
@@ -11,14 +11,14 @@ import numpy as np
 import torch
 from flwr.client import Client, NumPyClient
 from omegaconf import OmegaConf, open_dict
-from torch.utils.data import DataLoader
 
 from .client import FJORD_CONFIG_TYPE, FjORDClient, get_agg_config
 from .dataset import load_data
-from .models import get_net, test
+from .models import get_net
+from .server import get_eval_fn
 from .strategy import FjORDFedAVG
 from .utils.logger import Logger
-from .utils.utils import get_parameters, save_model, set_parameters
+from .utils.utils import get_parameters
 
 
 def get_fit_config_fn(
@@ -40,46 +40,6 @@ def get_fit_config_fn(
         return config
 
     return fit_config
-
-
-def get_eval_fn(
-    args: Any, model_path: str, testloader: DataLoader, device: torch.device
-):
-    """Get evaluation function.
-
-    :param args: Arguments
-    :param model_path: Path to save the model
-    :param testloader: Test data loader
-    :param device: Device to be used
-    :return: Evaluation function
-    """
-
-    def evaluate(
-        server_round: int,
-        parameters: fl.common.NDArrays,
-        config: Dict[str, fl.common.Scalar],  # pylint: disable=unused-argument
-    ) -> Optional[Tuple[float, Dict[str, fl.common.Scalar]]]:
-        if server_round and (server_round % args.evaluate_every == 0):
-            net = get_net(args.model, args.p_s, device)
-            set_parameters(net, parameters)
-            # Update model with the latest parameters
-            losses, accuracies = test(net, testloader, args.p_s)
-            avg_loss = sum(losses) / len(losses)
-            for p, loss, accuracy in zip(args.p_s, losses, accuracies):
-                Logger.get().info(
-                    f"Server-side evaluation (global round={server_round})"
-                    f" {p=}: {loss=} / {accuracy=}"
-                )
-            save_model(net, model_path)
-
-            return avg_loss, {
-                f"Accuracy[{p}]": acc for p, acc in zip(args.p_s, accuracies)
-            }
-
-        Logger.get().debug(f"Evaluation skipped for global round={server_round}.")
-        return float("inf"), {"accuracy": "None"}
-
-    return evaluate
 
 
 def get_client_fn(  # pylint: disable=too-many-arguments
